Refactor some code parts, cleanup added new endpoints
