+++ client/src/actions/navigations.js
@@ -1,0 +1,20 @@
+export const navKnowledgeBase = async () => {
+  const navigation = await fetch(`/api/navigations/knowledge`);
+  const result = await navigation.json();
+  console.error(result);
+  return result;
+};
+
+export const navCourses = async () => {
+  const navigation = await fetch(`/api/navigations/courses`);
+  const result = await navigation.json();
+  console.error(result);
+  return result;
+};
+
+export const navContent = async () => {
+  const navigation = await fetch(`/api/navigations/content`);
+  const result = await navigation.json();
+  console.error(result);
+  return result;
+};
